# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise AI Platform API
  version: 1.0.0
  description: >-
    Phase-one contract for enterprise knowledge search, document production,
    visual workflow execution, assets and traces. Provider credentials never
    appear in requests or responses.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
  /v1/knowledge-bases:
    get:
      operationId: listKnowledgeBases
      responses:
        "200":
          description: Visible knowledge bases after server-side authorization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./schemas/knowledge.schema.json#/$defs/knowledgeBase
    post:
      operationId: createKnowledgeBase
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, ownerDepartmentId, securityLevel, embeddingModelCode]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 1000 }
                ownerDepartmentId: { type: string }
                securityLevel:
                  $ref: ./schemas/common.schema.json#/$defs/securityLevel
                embeddingModelCode: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: ./schemas/knowledge.schema.json#/$defs/knowledgeBase
        default:
          $ref: "#/components/responses/Error"
  /v1/knowledge-bases/{knowledgeBaseId}/documents:
    post:
      operationId: uploadKnowledgeDocument
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                dataScope:
                  $ref: ./schemas/common.schema.json#/$defs/dataScope
                securityLevel:
                  $ref: ./schemas/common.schema.json#/$defs/securityLevel
                projectId: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "202":
          description: Accepted for parsing and indexing
          content:
            application/json:
              schema:
                $ref: ./schemas/knowledge.schema.json#/$defs/document
        default:
          $ref: "#/components/responses/Error"
  /v1/knowledge-bases/{knowledgeBaseId}/search:
    post:
      operationId: searchKnowledge
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/knowledge.schema.json#/$defs/searchRequest
      responses:
        "200":
          description: Permission-filtered citations
          content:
            application/json:
              schema:
                $ref: ./schemas/knowledge.schema.json#/$defs/searchResponse
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
    KnowledgeBaseId:
      name: knowledgeBaseId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Stable machine-readable error
      content:
        application/problem+json:
          schema:
            $ref: ./schemas/common.schema.json#/$defs/problemDetails